introduce card to typescript
