+++ package.json
@@ -48,6 +48,7 @@
     "typescript": "4.2.3"
   },
   "devDependencies": {
+    "@types/chart.js": "^2.9.34",
     "@types/react": "^17.0.19",
     "@types/react-dom": "^17.0.9",
     "@types/react-router-dom": "^5.1.8"
