+++ src/components/Cards/CardBarChart.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import Chart from "chart";
+import Chart from "chart.js";
 
-export default function CardLineChart() {
+export default function CardBarChart() {
   React.useEffect(() => {
-    var config = {
-      type: "line",
+    let config: any = {
+      type: "bar",
       data: {
         labels: [
           "January",
@@ -18,17 +18,19 @@ export default function CardLineChart() {
         datasets: [
           {
             label: new Date().getFullYear(),
-            backgroundColor: "#4c51bf",
-            borderColor: "#4c51bf",
-            data: [65, 78, 66, 44, 56, 67, 75],
+            backgroundColor: "#ed64a6",
+            borderColor: "#ed64a6",
+            data: [30, 78, 56, 34, 100, 45, 13],
             fill: false,
+            barThickness: 8,
           },
           {
             label: new Date().getFullYear() - 1,
             fill: false,
-            backgroundColor: "#fff",
-            borderColor: "#fff",
-            data: [40, 68, 86, 74, 56, 60, 87],
+            backgroundColor: "#4c51bf",
+            borderColor: "#4c51bf",
+            data: [27, 68, 86, 74, 10, 4, 87],
+            barThickness: 8,
           },
         ],
       },
@@ -37,15 +39,7 @@ export default function CardLineChart() {
         responsive: true,
         title: {
           display: false,
-          text: "Sales Charts",
-          fontColor: "white",
-        },
-        legend: {
-          labels: {
-            fontColor: "white",
-          },
-          align: "end",
-          position: "bottom",
+          text: "Orders Chart",
         },
         tooltips: {
           mode: "index",
@@ -55,24 +49,26 @@ export default function CardLineChart() {
           mode: "nearest",
           intersect: true,
         },
+        legend: {
+          labels: {
+            fontColor: "rgba(0,0,0,.4)",
+          },
+          align: "end",
+          position: "bottom",
+        },
         scales: {
           xAxes: [
             {
-              ticks: {
-                fontColor: "rgba(255,255,255,.7)",
-              },
-              display: true,
+              display: false,
               scaleLabel: {
-                display: false,
+                display: true,
                 labelString: "Month",
-                fontColor: "white",
               },
               gridLines: {
-                display: false,
                 borderDash: [2],
                 borderDashOffset: [2],
                 color: "rgba(33, 37, 41, 0.3)",
-                zeroLineColor: "rgba(0, 0, 0, 0)",
+                zeroLineColor: "rgba(33, 37, 41, 0.3)",
                 zeroLineBorderDash: [2],
                 zeroLineBorderDashOffset: [2],
               },
@@ -80,21 +76,17 @@ export default function CardLineChart() {
           ],
           yAxes: [
             {
-              ticks: {
-                fontColor: "rgba(255,255,255,.7)",
-              },
               display: true,
               scaleLabel: {
                 display: false,
                 labelString: "Value",
-                fontColor: "white",
               },
               gridLines: {
-                borderDash: [3],
-                borderDashOffset: [3],
+                borderDash: [2],
                 drawBorder: false,
-                color: "rgba(255, 255, 255, 0.15)",
-                zeroLineColor: "rgba(33, 37, 41, 0)",
+                borderDashOffset: [2],
+                color: "rgba(33, 37, 41, 0.2)",
+                zeroLineColor: "rgba(33, 37, 41, 0.15)",
                 zeroLineBorderDash: [2],
                 zeroLineBorderDashOffset: [2],
               },
@@ -103,26 +95,29 @@ export default function CardLineChart() {
         },
       },
     };
-    var ctx = document.getElementById("line-chart").getContext("2d");
-    window.myLine = new Chart(ctx, config);
+    let ctx = (document.getElementById("bar-chart") as HTMLCanvasElement)!.getContext("2d");
+    if(ctx)
+      window.myBar = new Chart(ctx, config);
   }, []);
   return (
     <>
-      <div className="relative flex flex-col min-w-0 break-words w-full mb-6 shadow-lg rounded bg-blueGray-700">
+      <div className="relative flex flex-col min-w-0 break-words bg-white w-full mb-6 shadow-lg rounded">
         <div className="rounded-t mb-0 px-4 py-3 bg-transparent">
           <div className="flex flex-wrap items-center">
             <div className="relative w-full max-w-full flex-grow flex-1">
-              <h6 className="uppercase text-blueGray-100 mb-1 text-xs font-semibold">
-                Overview
+              <h6 className="uppercase text-blueGray-400 mb-1 text-xs font-semibold">
+                Performance
               </h6>
-              <h2 className="text-white text-xl font-semibold">Sales value</h2>
+              <h2 className="text-blueGray-700 text-xl font-semibold">
+                Total orders
+              </h2>
             </div>
           </div>
         </div>
         <div className="p-4 flex-auto">
           {/* Chart */}
           <div className="relative h-350-px">
-            <canvas id="line-chart"></canvas>
+            <canvas id="bar-chart"></canvas>
           </div>
         </div>
       </div>
