+++ src/components/Cards/CardLineChart.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import Chart from "chart";
+import Chart from "chart.js";
 
-export default function CardBarChart() {
+export default function CardLineChart() {
   React.useEffect(() => {
-    let config = {
-      type: "bar",
+    let config: any = {
+      type: "line",
       data: {
         labels: [
           "January",
@@ -18,19 +18,17 @@ export default function CardBarChart() {
         datasets: [
           {
             label: new Date().getFullYear(),
-            backgroundColor: "#ed64a6",
-            borderColor: "#ed64a6",
-            data: [30, 78, 56, 34, 100, 45, 13],
+            backgroundColor: "#4c51bf",
+            borderColor: "#4c51bf",
+            data: [65, 78, 66, 44, 56, 67, 75],
             fill: false,
-            barThickness: 8,
           },
           {
             label: new Date().getFullYear() - 1,
             fill: false,
-            backgroundColor: "#4c51bf",
-            borderColor: "#4c51bf",
-            data: [27, 68, 86, 74, 10, 4, 87],
-            barThickness: 8,
+            backgroundColor: "#fff",
+            borderColor: "#fff",
+            data: [40, 68, 86, 74, 56, 60, 87],
           },
         ],
       },
@@ -39,7 +37,15 @@ export default function CardBarChart() {
         responsive: true,
         title: {
           display: false,
-          text: "Orders Chart",
+          text: "Sales Charts",
+          fontColor: "white",
+        },
+        legend: {
+          labels: {
+            fontColor: "white",
+          },
+          align: "end",
+          position: "bottom",
         },
         tooltips: {
           mode: "index",
@@ -49,26 +55,24 @@ export default function CardBarChart() {
           mode: "nearest",
           intersect: true,
         },
-        legend: {
-          labels: {
-            fontColor: "rgba(0,0,0,.4)",
-          },
-          align: "end",
-          position: "bottom",
-        },
         scales: {
           xAxes: [
             {
-              display: false,
+              ticks: {
+                fontColor: "rgba(255,255,255,.7)",
+              },
+              display: true,
               scaleLabel: {
-                display: true,
+                display: false,
                 labelString: "Month",
+                fontColor: "white",
               },
               gridLines: {
+                display: false,
                 borderDash: [2],
                 borderDashOffset: [2],
                 color: "rgba(33, 37, 41, 0.3)",
-                zeroLineColor: "rgba(33, 37, 41, 0.3)",
+                zeroLineColor: "rgba(0, 0, 0, 0)",
                 zeroLineBorderDash: [2],
                 zeroLineBorderDashOffset: [2],
               },
@@ -76,17 +80,21 @@ export default function CardBarChart() {
           ],
           yAxes: [
             {
+              ticks: {
+                fontColor: "rgba(255,255,255,.7)",
+              },
               display: true,
               scaleLabel: {
                 display: false,
                 labelString: "Value",
+                fontColor: "white",
               },
               gridLines: {
-                borderDash: [2],
+                borderDash: [3],
+                borderDashOffset: [3],
                 drawBorder: false,
-                borderDashOffset: [2],
-                color: "rgba(33, 37, 41, 0.2)",
-                zeroLineColor: "rgba(33, 37, 41, 0.15)",
+                color: "rgba(255, 255, 255, 0.15)",
+                zeroLineColor: "rgba(33, 37, 41, 0)",
                 zeroLineBorderDash: [2],
                 zeroLineBorderDashOffset: [2],
               },
@@ -95,28 +103,27 @@ export default function CardBarChart() {
         },
       },
     };
-    let ctx = document.getElementById("bar-chart").getContext("2d");
-    window.myBar = new Chart(ctx, config);
+    var ctx = (document.getElementById("line-chart") as HTMLCanvasElement)!.getContext("2d");
+    if(ctx)
+      window.myLine = new Chart(ctx, config);
   }, []);
   return (
     <>
-      <div className="relative flex flex-col min-w-0 break-words bg-white w-full mb-6 shadow-lg rounded">
+      <div className="relative flex flex-col min-w-0 break-words w-full mb-6 shadow-lg rounded bg-blueGray-700">
         <div className="rounded-t mb-0 px-4 py-3 bg-transparent">
           <div className="flex flex-wrap items-center">
             <div className="relative w-full max-w-full flex-grow flex-1">
-              <h6 className="uppercase text-blueGray-400 mb-1 text-xs font-semibold">
-                Performance
+              <h6 className="uppercase text-blueGray-100 mb-1 text-xs font-semibold">
+                Overview
               </h6>
-              <h2 className="text-blueGray-700 text-xl font-semibold">
-                Total orders
-              </h2>
+              <h2 className="text-white text-xl font-semibold">Sales value</h2>
             </div>
           </div>
         </div>
         <div className="p-4 flex-auto">
           {/* Chart */}
           <div className="relative h-350-px">
-            <canvas id="bar-chart"></canvas>
+            <canvas id="line-chart"></canvas>
           </div>
         </div>
       </div>
