+++ src/components/Cards/CardProfile.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 // components
 
 export default function CardProfile() {
