+++ src/components/Cards/CardSettings.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 // components
 
 export default function CardSettings() {
@@ -168,10 +166,9 @@ export default function CardSettings() {
                     About me
                   </label>
                   <textarea
-                    type="text"
                     className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
                     defaultValue="A beautiful UI Kit and Admin for React & Tailwind CSS. It is Free and Open Source."
-                    rows="4"
+                    rows={4}
                   ></textarea>
                 </div>
               </div>
